Rename files, add modal close on dismount
